Cleanup

+++ src/delay.h
@@ -6,7 +6,8 @@
 
 void _delay_us(uint32_t d);
 void _delay_ms(uint32_t d);
-#define TICKS_PER_MS (F_CPU/1000)
-#define TICKS_PER_US (TICKS_PER_MS/1000)
+
+#define TICKS_PER_MS ((uint32_t) (F_CPU/1000.0))
+#define TICKS_PER_US ((uint32_t) (TICKS_PER_MS/1000.0))
 
 #endif // _DELAY_H
